Only admins can create, edit, or delete projects
