+++ app/controllers/projects_controller.rb
@@ -7,42 +7,7 @@ class ProjectsController < ApplicationController
     @projects = Project.all
   end
 
-  def new
-    @project = Project.new
-  end
-
-  def create
-    @project = Project.new(project_params)
-
-    if @project.save
-      flash[:notice] = 'Project has been created.'
-      redirect_to @project
-    else
-      flash.now[:alert] = 'Project has not been created.'
-      render 'new'
-    end
-  end
-
   def show; end
-
-  def edit; end
-
-  def update
-    if @project.update(project_params)
-      flash[:notice] = 'Project has been updated.'
-      redirect_to @project
-    else
-      flash.now[:alert] = 'Project has not been updated.'
-      render 'edit'
-    end
-  end
-
-  def destroy
-    return unless @project.delete
-
-    flash[:notice] = 'Project has been deleted.'
-    redirect_to projects_path
-  end
 
   private
