+++ app/helpers/admin/projects_helper.rb
@@ -1,0 +1,2 @@
+module Admin::ProjectsHelper
+end
