+++ app/views/admin/projects/_form.html.erb
@@ -1,0 +1,5 @@
+<%= bootstrap_form_with(model: [:admin, project], local: true, label_errors: true) do |form| %>
+  <%= form.text_field :name %>
+  <%= form.text_field :description %>
+  <%= form.primary %>
+<% end %>
