--- app/views/projects/_form.html.erb
@@ -1,6 +1,0 @@
-<%= bootstrap_form_with(model: project, local: true, label_errors: true) do |form| %>
-  <%= form.text_field :name %>
-  <%= form.text_field :description %>
-
-  <%= form.primary %>
-<% end %>
