+++ app/views/projects/index.html.erb
@@ -1,11 +1,9 @@
 <header>
   <h1>Projects</h1>
-
   <ul class="actions">
-    <li><%= link_to "New Project", new_project_path, class: 'new' %></li>
+    <li><%= link_to "New Project", new_admin_project_path, class: 'new' %></li>
   </ul>
 </header>
-
 <div class="projects">
   <% @projects.each do |project| %>
     <h2><%= link_to project.name, project %></h2>
