+++ app/views/projects/show.html.erb
@@ -1,24 +1,19 @@
 <% title @project.name, 'Projects' %>
-
 <header>
   <h1><%= @project.name %></h1>
-
   <ul class="actions">
-    <li><%= link_to 'Edit Project', edit_project_path(@project),
+    <li><%= link_to 'Edit Project', edit_admin_project_path(@project),
       class: 'edit' %></li>
     <li><%= link_to 'Delete Project',
-    project_path(@project),
+    admin_project_path(@project),
     method: :delete,
     data: { confirm: "Are you sure you want to delete this project?" },
     class: "delete" %></li>
   </ul>
 </header>
-
 <p><%= @project.description %></p>
-
 <header class="page-header">
   <h2>Tickets</h2>
-
   <ul class="actions">
     <li>
       <%= link_to 'New Ticket',
@@ -27,7 +22,6 @@
     </li>
   </ul>
 </header>
-
 <ul class="tickets">
   <% @project.tickets.each do |ticket| %>
     <li>
