+++ spec/features/admin/deleting_projects_spec.rb
@@ -3,6 +3,10 @@
 require 'rails_helper'
 
 RSpec.feature 'Users can delete projects' do
+  before do
+    login_as(FactoryBot.create(:user, :admin))
+  end
+
   scenario 'successfully' do
     FactoryBot.create(:project, name: 'Visual Studio Code')
