+++ spec/features/admin/editing_projects_spec.rb
@@ -4,6 +4,7 @@ require 'rails_helper'
 
 RSpec.feature 'Users can edit existing projects' do
   before do
+    login_as(FactoryBot.create(:user, :admin))
     FactoryBot.create(:project, name: 'Visual Studio Code')
 
     visit '/'
