+++ spec/requests/admin/projects_request_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe "Admin::Projects", type: :request do
+
+end
